+++ lib/roda/data_grid/dataminer_control.rb
@@ -344,14 +344,14 @@ class DataminerControl # rubocop:disable Metrics/ClassLength
     path = File.join(@root, 'grid_definitions', 'searches', file_name.sub('.yml', '') << '.yml')
     YAML.load(File.read(path))
   rescue Psych::SyntaxError => e
-    raise "Syntax error in YAML file (#{file_name.sub('.yml', '') << '.yml'}). The syntax error is: #{e.message}"
+    raise Error, "Syntax error in YAML file (#{file_name.sub('.yml', '') << '.yml'}). The syntax error is: #{e.message}"
   end
 
   def load_list_definition(file_name)
     path = File.join(@root, 'grid_definitions', 'lists', file_name.sub('.yml', '') << '.yml')
     YAML.load(File.read(path))
   rescue Psych::SyntaxError => e
-    raise "Syntax error in YAML file (#{file_name.sub('.yml', '') << '.yml'}). The syntax error is: #{e.message}"
+    raise Error, "Syntax error in YAML file (#{file_name.sub('.yml', '') << '.yml'}). The syntax error is: #{e.message}"
   end
 
   # Load a YML report.